BUGZID:20915 Fix the problem that weaver crashs every time when webconfig can not find any image file.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/wci18n@11246 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/wvlinklist.h
@@ -310,8 +310,10 @@ public:
      * 
      */ 
     void unlink_first()
+    { 
+        if(head.next != NULL)   
         { Iter i(*this); i.rewind(); i.next(); i.unlink(); }
-
+    }
     /**
      * Unlinks the element that follows the specified link in the list.
      * 
@@ -323,12 +325,15 @@ public:
     void unlink_after(WvLink *after, bool destroy = true)
     {
         WvLink *next = after->next;
-        T *obj = (destroy && next->get_autofree()) ?
+        if(next != NULL)
+        {
+            T *obj = (destroy && next->get_autofree()) ?
             static_cast<T*>(next->data) : NULL;
-        if (next == tail) tail = after;
-        next->unlink(after);
-	if (obj)
-	    WvTraits<T>::release(obj);
+            if (next == tail) tail = after;
+            next->unlink(after);
+	    if (obj)
+	        WvTraits<T>::release(obj);
+        }
     }
 
     /**
